requirements.txt
pydantic
python-dotenv
langchain
langchain-core
langchain-community
langchain-openai
fastapi
uvicorn[standard]
httpx
